implemented dragon interface
